roughly working demo of PINNs

+++ archive/ml-pinn/src/base/dtype.py
@@ -2,6 +2,6 @@ __all__ = ["DTYPE"]
 
 import tensorflow as tf
 
-# DTYPE='float32' # default
-DTYPE='float64' # for higher-order derivatives
+DTYPE='float32' # default
+# DTYPE='float64' # for higher-order derivatives
 tf.keras.backend.set_floatx(DTYPE)
